pyproject.toml
[tool.poetry]
name = "prayer-room-api"
version = "0.1.0"
description = ""
authors = ["Andrew Miller <info@akmiller.co.uk>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
django = "^5.1.5"
django-rest-framework = "^0.1.0"
django-import-export = "^4.3.4"
python-dateutil = "^2.9.0.post0"
django-cors-headers = "^4.6.0"
django-webhook = "^0.0.15"
celery = "^5.4.0"
django-allauth = "^65.3.1"
gunicorn = "^23.0.0"
whitenoise = "^6.8.2"
psycopg = "^3.2.4"
django-classy-settings = "^3.0.7"
python-dotenv = "^1.0.1"
sentry-sdk = { extras = ["django"], version = "^2.22.0" }
django-extensions = "^4.1"
django-htmx = "^1.21"
typing-extensions = "^4.12.2"
neapolitan = "^25.1"
django-anymail = { extras = ["amazon-ses"], version = "^12.0" }
markdown = "^3.7"
django-prodserver = "^2.4.0"
django-celery-beat = "^2.8.1"


[tool.poetry.group.dev.dependencies]
django-debug-toolbar = "^5.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
